+++ packages/tests/@realtimetests/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @realtimetests/react
 
+## 0.1.1
+
+### Patch Changes
+
+- Test splitting group and document context
+
 ## 0.1.0
 
 ### Minor Changes